--- generators/app/templates/src/app/screens/Registration copy/constants.tsx
@@ -1,3 +1,0 @@
-import { stringArrayToObject } from '~utils/array';
-
-export const FIELDS = stringArrayToObject(['name', 'surname', 'email', 'password', 'confirmPassword']);
