--- generators/app/templates/src/app/screens/Registration copy/i18n.ts
@@ -1,18 +1,0 @@
-import i18next from 'i18next';
-
-i18next.addResources('es', 'Registration', {
-  registration: 'Creación de cuenta',
-  registrationExplanation: 'Ingrese los siguientes datos para crear su cuenta',
-  name: 'Nombre',
-  namePlaceholder: 'Ingrese su nombre',
-  surname: 'Apellido',
-  surnamePlaceholder: 'Ingrese su apellido',
-  email: 'E-mail',
-  emailPlaceholder: 'Ingrese su e-mail',
-  password: 'Contraseña',
-  passwordPlaceholder: 'Ingrese su contraseña',
-  confirmPassword: 'Confirmar contraseña',
-  confirmPasswordPlaceholder: 'Vuelva a ingresar su contraseña',
-  required: '* Requerido',
-  submit: 'Enviar'
-});
