--- generators/app/templates/src/app/screens/Registration copy/styles.module.scss
@@ -1,22 +1,0 @@
-.button {
-  border: solid 1px;
-  border-radius: 5px;
-  height: 40px;
-}
-
-.input {
-  height: 30px;
-  border: solid 1px;
-  border-radius: 5px;
-  padding: 5px;
-  font-size: 14px;
-}
-
-.form-container {
-  min-height: 500px;
-}
-
-.section-container {
-  width: 100%;
-  max-width: 450px;
-}
