+++ generators/app/templates/src/app/screens/Registration/constants.tsx
@@ -1,3 +1,3 @@
-import { stringArrayToObject } from '../../../utils/array';
+import { stringArrayToObject } from '~utils/array';
 
 export const FIELDS = stringArrayToObject(['name', 'surname', 'email', 'password', 'confirmPassword']);
