+++ generators/app/templates/src/app/screens/Registration/index.test.js
