+++ generators/app/templates/src/utils/array.ts
@@ -13,7 +13,7 @@ export function stringArrayToObject(actionsArray: string[], namespace = ''): { [
     throw new Error('Action names must be strings and must not be empty');
   }
 
-  return new Immutable(actionsArray).asObject((actionName: string) => [
+  return new (Immutable as any)(actionsArray).asObject((actionName: string) => [
     actionName,
     `${namespace}:${actionName}`
   ]);
